feat: Support integrations with generic prefix.
Some integrations do not use homeassistant.component. prefix (eg. totp)
and we need to enable them to bootstrap HA core. Instead of just
supporting the homeassistant.component. prefix, support all prefixes.

If the supplied integration doesn't have a '.' in the name, we treat
this implicitly as an integration with `homeassistant.component.`
prefix.